Raise NotImplemented error
This file is not meant to be run for now. The code depends on scipy, which is not
added to the pyproject.toml file. Also, the need for this file is not clear, since we
might not implement a music feature. If you want to run this file, you'll need to
add it with poetry as an dependency.